Auto-skip: add support for arg matrices
