+++ inputgraph/util.go
@@ -1,21 +1,5 @@
 package inputgraph
 
-import (
-	"strings"
-)
-
-func requiresCrossProduct(args []string) bool {
-	seen := map[string]struct{}{}
-	for _, s := range args {
-		k := strings.SplitN(s, "=", 2)[0]
-		if _, found := seen[k]; found {
-			return true
-		}
-		seen[k] = struct{}{}
-	}
-	return false
-}
-
 func copyVisited(m map[string]struct{}) map[string]struct{} {
 	m2 := map[string]struct{}{}
 	for k := range m {
